refactor to combine azure blob and s3 tests
